Move Kafka implementation to deprecated
