+++ tensorflow_io/core/kernels/kafka_kernels_deprecated.cc
@@ -17,9 +17,18 @@ limitations under the License.
 #include "tensorflow/core/framework/resource_mgr.h"
 #include "tensorflow/core/framework/resource_op_kernel.h"
 #include "tensorflow/core/platform/logging.h"
+#include "tensorflow_io/core/kernels/io_interface.h"
 #include "tensorflow_io/core/kernels/sequence_ops.h"
 
+#include "api/Compiler.hh"
+#include "api/DataFile.hh"
+#include "api/Generic.hh"
+#include "api/Stream.hh"
+#include "api/Validator.hh"
+#include "rdkafkacpp.h"
+
 #include <deque>
+#include <unordered_map>
 #include "rdkafkacpp.h"
 
 namespace tensorflow {
@@ -737,6 +746,631 @@ class KafkaOutputSequenceOp : public OutputSequenceOp<KafkaOutputSequence> {
   }
 };
 
+namespace data {
+
+class KafkaEventCb : public RdKafka::EventCb {
+ public:
+  KafkaEventCb() : run_(true) {}
+
+  bool run() { return run_; }
+
+  void event_cb(RdKafka::Event& event) {
+    switch (event.type()) {
+      case RdKafka::Event::EVENT_ERROR:
+        LOG(ERROR) << "EVENT_ERROR: "
+                   << "(" << RdKafka::err2str(event.err())
+                   << "): " << event.str();
+        { run_ = (event.err() != RdKafka::ERR__ALL_BROKERS_DOWN); }
+        break;
+      case RdKafka::Event::EVENT_STATS:
+        LOG(ERROR) << "EVENT_STATS: " << event.str();
+        break;
+      case RdKafka::Event::EVENT_LOG:
+        LOG(ERROR) << "EVENT_LOG: " << event.severity() << "-"
+                   << event.fac().c_str() << "-" << event.str().c_str();
+        break;
+      case RdKafka::Event::EVENT_THROTTLE:
+        LOG(ERROR) << "EVENT_THROTTLE: " << event.throttle_time() << "ms by "
+                   << event.broker_name() << " id " << (int)event.broker_id();
+        break;
+      default:
+        LOG(ERROR) << "EVENT: " << event.type() << " ("
+                   << RdKafka::err2str(event.err()) << "): " << event.str();
+        break;
+    }
+  }
+
+ private:
+  mutable mutex mu_;
+  bool run_ GUARDED_BY(mu_) = true;
+};
+
+class KafkaReadable : public IOReadableInterface {
+ public:
+  KafkaReadable(Env* env) : env_(env) {}
+
+  ~KafkaReadable() {}
+  Status Init(const std::vector<string>& input,
+              const std::vector<string>& metadata, const void* memory_data,
+              const int64 memory_size) override {
+    std::unique_ptr<RdKafka::Conf> conf(
+        RdKafka::Conf::create(RdKafka::Conf::CONF_GLOBAL));
+    std::unique_ptr<RdKafka::Conf> conf_topic(
+        RdKafka::Conf::create(RdKafka::Conf::CONF_TOPIC));
+
+    string errstr;
+    RdKafka::Conf::ConfResult result = RdKafka::Conf::CONF_UNKNOWN;
+
+    eof_ = true;
+    timeout_ = 2000;
+    for (size_t i = 0; i < metadata.size(); i++) {
+      if (metadata[i].find("conf.eof") == 0) {
+        std::vector<string> parts = str_util::Split(metadata[i], "=");
+        if (parts.size() != 2) {
+          return errors::InvalidArgument("invalid bounded configuration: ",
+                                         metadata[i]);
+        }
+        eof_ = (parts[1] != "0");
+      } else if (metadata[i].find("conf.timeout") == 0) {
+        std::vector<string> parts = str_util::Split(metadata[i], "=");
+        if (parts.size() != 2 || !strings::safe_strto64(parts[1], &timeout_)) {
+          return errors::InvalidArgument("invalid timeout configuration: ",
+                                         metadata[i]);
+        }
+      } else if (metadata[i].find("conf.topic.") == 0) {
+        std::vector<string> parts = str_util::Split(metadata[i], "=");
+        if (parts.size() != 2) {
+          return errors::InvalidArgument("invalid topic configuration: ",
+                                         metadata[i]);
+        }
+        result = conf_topic->set(parts[0].substr(11), parts[1], errstr);
+        if (result != RdKafka::Conf::CONF_OK) {
+          return errors::Internal("failed to do topic configuration:",
+                                  metadata[i], "error:", errstr);
+        }
+      } else if (metadata[i] != "" &&
+                 metadata[i].find("conf.") == string::npos) {
+        std::vector<string> parts = str_util::Split(metadata[i], "=");
+        if (parts.size() != 2) {
+          return errors::InvalidArgument("invalid topic configuration: ",
+                                         metadata[i]);
+        }
+        if ((result = conf->set(parts[0], parts[1], errstr)) !=
+            RdKafka::Conf::CONF_OK) {
+          return errors::Internal("failed to do global configuration: ",
+                                  metadata[i], "error:", errstr);
+        }
+      }
+      LOG(INFO) << "Kafka configuration: " << metadata[i];
+    }
+    if ((result = conf->set("default_topic_conf", conf_topic.get(), errstr)) !=
+        RdKafka::Conf::CONF_OK) {
+      return errors::Internal("failed to set default_topic_conf:", errstr);
+    }
+
+    // consumer.properties:
+    //   bootstrap.servers=localhost:9092
+    //   group.id=test-consumer-group
+    string bootstrap_servers;
+    if ((result = conf->get("bootstrap.servers", bootstrap_servers)) !=
+        RdKafka::Conf::CONF_OK) {
+      bootstrap_servers = "localhost:9092";
+      if ((result = conf->set("bootstrap.servers", bootstrap_servers,
+                              errstr)) != RdKafka::Conf::CONF_OK) {
+        return errors::Internal("failed to set bootstrap.servers [",
+                                bootstrap_servers, "]:", errstr);
+      }
+    }
+    string group_id;
+    if ((result = conf->get("group.id", group_id)) != RdKafka::Conf::CONF_OK) {
+      group_id = "test-consumer-group";
+      if ((result = conf->set("group.id", group_id, errstr)) !=
+          RdKafka::Conf::CONF_OK) {
+        return errors::Internal("failed to set group.id [", group_id,
+                                "]:", errstr);
+      }
+    }
+    if ((result = conf->set("event_cb", &kafka_event_cb_, errstr)) !=
+        RdKafka::Conf::CONF_OK) {
+      return errors::Internal("failed to set event_cb:", errstr);
+    }
+
+    // Always enable.partition.eof=true
+    if ((result = conf->set("enable.partition.eof", "true", errstr)) !=
+        RdKafka::Conf::CONF_OK) {
+      return errors::Internal("Failed to set enable.partition.eof=true :",
+                              errstr);
+    }
+
+    consumer_.reset(RdKafka::KafkaConsumer::create(conf.get(), errstr));
+    if (!consumer_.get()) {
+      return errors::Internal("failed to create consumer:", errstr);
+    }
+
+    // TODO: multiple topic and partitions
+    const string& entry = input[0];
+    std::vector<string> parts = str_util::Split(entry, ":");
+    if (parts.size() != 3 && parts.size() != 4) {
+      return errors::InvalidArgument("invalid input: ", entry);
+    }
+    string topic = parts[0];
+    int32 partition = 0;
+    if (!strings::safe_strto32(parts[1], &partition)) {
+      return errors::InvalidArgument("invalid parameters: ", entry);
+    }
+    subscription_.reset(RdKafka::TopicPartition::create(topic, partition));
+    std::vector<RdKafka::TopicPartition*> partitions;
+    partitions.emplace_back(subscription_.get());
+
+    offset_head_ = 0;
+    offset_tail_ = -1;
+    if (!strings::safe_strto64(parts[2], &offset_head_)) {
+      return errors::InvalidArgument("invalid parameters: ", entry);
+    }
+    subscription_->set_offset(offset_head_);
+    RdKafka::ErrorCode err = consumer_->assign(partitions);
+    if (err != RdKafka::ERR_NO_ERROR) {
+      return errors::Internal("failed to assign partition: ",
+                              RdKafka::err2str(err));
+    }
+    // If offset tail is specified then resolve
+    if (parts.size() == 4) {
+      std::unique_ptr<RdKafka::Message> message;
+      do {
+        message.reset(consumer_->consume(timeout_));
+      } while (message->err() == RdKafka::ERR__TRANSPORT);
+      if (message->err() != RdKafka::ERR_NO_ERROR) {
+        return errors::Internal("failed to consume head message: ",
+                                RdKafka::err2str(message->err()));
+      }
+      offset_head_ = message->offset();
+
+      if (!strings::safe_strto64(parts[3], &offset_tail_)) {
+        return errors::InvalidArgument("invalid parameters: ", entry);
+      }
+      if (offset_tail_ == -1) {
+        subscription_->set_offset(RdKafka::Consumer::OffsetTail(1));
+      } else {
+        subscription_->set_offset(offset_tail_ - 1);
+      }
+      err = consumer_->seek(*subscription_, timeout_);
+      if (err != RdKafka::ERR_NO_ERROR) {
+        return errors::Internal("failed to seek tail -1: ",
+                                RdKafka::err2str(err));
+      }
+      do {
+        message.reset(consumer_->consume(timeout_));
+      } while (message->err() == RdKafka::ERR__TRANSPORT);
+      if (message->err() != RdKafka::ERR_NO_ERROR) {
+        return errors::Internal("failed to consume tail message: ",
+                                RdKafka::err2str(message->err()));
+      }
+      offset_tail_ = message->offset();
+
+      subscription_->set_offset(offset_head_);
+      err = consumer_->seek(*subscription_, timeout_);
+      if (err != RdKafka::ERR_NO_ERROR) {
+        return errors::Internal("failed to seek tail -1: ",
+                                RdKafka::err2str(err));
+      }
+    }
+
+    index_ = 0;
+    offset_ = subscription_->offset();
+    return Status::OK();
+  }
+  Status Read(const int64 start, const int64 stop, const string& component,
+              int64* record_read, Tensor* value, Tensor* label) override {
+    *record_read = 0;
+    if (start != index_) {
+      // Case 1: start > 0
+      if (start > 0) {
+        // If EOF has aleady been reached, then we will just return empty
+        // in case start > index_
+        if (start > index_) {
+          // EOF of topic for stream dataset, just return
+          if (consumer_.get() == nullptr) {
+            return Status::OK();
+          }
+          // EOF of topic for normal dataset, just return
+          if (offset_tail_ >= 0 && offset_ >= offset_tail_) {
+            return Status::OK();
+          }
+        }
+        // Otherwise  just return error
+        return errors::InvalidArgument(
+            "dataset can not seek to a random location");
+      }
+      // Case 2: start = 0
+      // If offset_tail_ < 0, then this is a stream dataset,
+      // we should return empty data
+      if (offset_tail_ < 0) {
+        return errors::InvalidArgument(
+            "stream dataset can not return back to 0");
+      }
+      // otherwise we seek to the beginning, this is important to make dataset
+      // repeatable so that it could be used in training
+      subscription_->set_offset(offset_head_);
+      RdKafka::ErrorCode err = consumer_->seek((*subscription_), timeout_);
+      if (err != RdKafka::ERR_NO_ERROR) {
+        return errors::Internal("failed to seek partition: ",
+                                RdKafka::err2str(err));
+      }
+      index_ = 0;
+      offset_ = subscription_->offset();
+      LOG(INFO) << "Kafka stream starts with current offset: "
+                << subscription_->offset();
+    }
+
+    while (consumer_.get() != nullptr && index_ < stop) {
+      if (!kafka_event_cb_.run()) {
+        return errors::Internal("failed to consume due to all brokers down");
+      }
+      if (offset_tail_ >= 0 && offset_ >= offset_tail_) {
+        // EOF of topic, reset index to 0
+        break;
+      }
+
+      std::unique_ptr<RdKafka::Message> message(consumer_->consume(timeout_));
+      if (message->err() == RdKafka::ERR_NO_ERROR) {
+        // Produce the line as output.
+        value->flat<string>()(index_) = std::string(
+            static_cast<const char*>(message->payload()), message->len());
+        (*record_read)++;
+        index_++;
+        offset_ = message->offset();
+        continue;
+      }
+      if (message->err() == RdKafka::ERR__PARTITION_EOF) {
+        LOG(INFO) << "Partition reach EOF, current offset: "
+                  << subscription_->offset();
+        if (offset_tail_ >= 0) {
+          return errors::Internal("received EOF unexpected: ",
+                                  message->errstr());
+        }
+        // If this is the end of file, then stop here.
+        // Also, this means, file could not be reused.
+        if (eof_) {
+          consumer_.reset(nullptr);
+          break;
+        }
+      } else if (message->err() == RdKafka::ERR__TRANSPORT) {
+        // Not return error here because consumer will try re-connect.
+        LOG(ERROR) << "Broker transport failure: " << message->errstr();
+      } else if (message->err() != RdKafka::ERR__TIMED_OUT) {
+        LOG(ERROR) << "Failed to consume: " << message->errstr();
+        return errors::Internal("Failed to consume: ", message->errstr());
+      }
+    }
+    return Status::OK();
+  }
+  Status Spec(const string& component, PartialTensorShape* shape,
+              DataType* dtype, bool label) override {
+    *shape = PartialTensorShape({-1});
+    *dtype = DT_STRING;
+    return Status::OK();
+  }
+
+  string DebugString() const override {
+    mutex_lock l(mu_);
+    return strings::StrCat("KafkaReadable[]");
+  }
+
+ private:
+  mutable mutex mu_;
+  Env* env_ GUARDED_BY(mu_);
+  int64 index_ GUARDED_BY(mu_);
+  int64 offset_ GUARDED_BY(mu_);
+
+  std::unique_ptr<RdKafka::TopicPartition> subscription_ GUARDED_BY(mu_);
+  std::unique_ptr<RdKafka::KafkaConsumer> consumer_ GUARDED_BY(mu_);
+  KafkaEventCb kafka_event_cb_ = KafkaEventCb();
+  int64 timeout_ GUARDED_BY(mu_);
+  bool eof_ GUARDED_BY(mu_);
+
+  int64 offset_head_ GUARDED_BY(mu_);
+  int64 offset_tail_ GUARDED_BY(mu_);
+};
+
+class DecodeAvroResource : public ResourceBase {
+ public:
+  DecodeAvroResource(Env* env) : env_(env) {}
+  ~DecodeAvroResource() {}
+
+  Status Init(const string& input) {
+    mutex_lock lock(mu_);
+    schema_ = input;
+    schema_stream_ = std::istringstream(schema_);
+
+    string error;
+    if (!(avro::compileJsonSchema(schema_stream_, avro_schema_, error))) {
+      return errors::Unimplemented("Avro schema error: ", error);
+    }
+
+    return Status::OK();
+  }
+  const avro::ValidSchema& avro_schema() { return avro_schema_; }
+  string DebugString() const override { return "DecodeAvroResource"; }
+
+ private:
+  mutable mutex mu_;
+  Env* env_ GUARDED_BY(mu_);
+  string schema_ GUARDED_BY(mu_);
+  std::istringstream schema_stream_;
+  avro::ValidSchema avro_schema_;
+};
+
+class DecodeAvroOp : public OpKernel {
+ public:
+  explicit DecodeAvroOp(OpKernelConstruction* context) : OpKernel(context) {
+    env_ = context->env();
+  }
+
+  void Compute(OpKernelContext* context) override {
+    const Tensor* input_tensor;
+    OP_REQUIRES_OK(context, context->input("input", &input_tensor));
+
+    DecodeAvroResource* resource;
+    std::unique_ptr<DecodeAvroResource> resource_scope;
+    if (context->input_dtype(1) == DT_RESOURCE) {
+      OP_REQUIRES_OK(context,
+                     GetResourceFromContext(context, "schema", &resource));
+    } else {
+      const Tensor* schema_tensor;
+      OP_REQUIRES_OK(context, context->input("schema", &schema_tensor));
+      const string& schema = schema_tensor->scalar<string>()();
+
+      resource_scope.reset(new DecodeAvroResource(env_));
+      OP_REQUIRES_OK(context, resource_scope->Init(schema));
+      resource_scope->Ref();
+      resource = resource_scope.get();
+    }
+    core::ScopedUnref unref(resource);
+
+    std::vector<Tensor*> value;
+    value.reserve(resource->avro_schema().root()->names());
+    for (size_t i = 0; i < resource->avro_schema().root()->names(); i++) {
+      Tensor* value_tensor = nullptr;
+      OP_REQUIRES_OK(context, context->allocate_output(static_cast<int64>(i),
+                                                       input_tensor->shape(),
+                                                       &value_tensor));
+      value.push_back(value_tensor);
+    }
+
+    avro::GenericDatum datum(resource->avro_schema());
+    for (int64 entry_index = 0; entry_index < input_tensor->NumElements();
+         entry_index++) {
+      const string& entry = input_tensor->flat<string>()(entry_index);
+      std::unique_ptr<avro::InputStream> in =
+          avro::memoryInputStream((const uint8_t*)entry.data(), entry.size());
+
+      avro::DecoderPtr d = avro::binaryDecoder();
+      d->init(*in);
+      avro::decode(*d, datum);
+      const avro::GenericRecord& record = datum.value<avro::GenericRecord>();
+      for (int i = 0; i < resource->avro_schema().root()->names(); i++) {
+        const avro::GenericDatum& field = record.fieldAt(i);
+        switch (field.type()) {
+          case avro::AVRO_NULL:
+            switch (context->expected_output_dtype(i)) {
+              case DT_BOOL:
+                value[i]->flat<bool>()(entry_index) = false;
+                break;
+              case DT_INT32:
+                value[i]->flat<int32>()(entry_index) = 0;
+                break;
+              case DT_INT64:
+                value[i]->flat<int64>()(entry_index) = 0;
+                break;
+              case DT_FLOAT:
+                value[i]->flat<float>()(entry_index) = 0.0;
+                break;
+              case DT_DOUBLE:
+                value[i]->flat<double>()(entry_index) = 0.0;
+                break;
+              case DT_STRING:
+                value[i]->flat<string>()(entry_index) = "";
+                break;
+              default:
+                OP_REQUIRES(context, false,
+                            errors::InvalidArgument(
+                                "unsupported data type against AVRO_NULL: ",
+                                field.type()));
+            }
+            break;
+          case avro::AVRO_BOOL:
+            value[i]->flat<bool>()(entry_index) = field.value<bool>();
+            break;
+          case avro::AVRO_INT:
+            value[i]->flat<int32>()(entry_index) = field.value<int32_t>();
+            break;
+          case avro::AVRO_LONG:
+            value[i]->flat<int64>()(entry_index) = field.value<int64_t>();
+            break;
+          case avro::AVRO_FLOAT:
+            value[i]->flat<float>()(entry_index) = field.value<float>();
+            break;
+          case avro::AVRO_DOUBLE:
+            value[i]->flat<double>()(entry_index) = field.value<double>();
+            break;
+          case avro::AVRO_STRING: {
+            // make a concrete explicit copy as otherwise avro may override the
+            // underlying buffer.
+            const string& field_value = field.value<string>();
+            string v;
+            if (field_value.size() > 0) {
+              v.resize(field_value.size());
+              memcpy(&v[0], &field_value[0], field_value.size());
+            }
+            value[i]->flat<string>()(entry_index) = v;
+          } break;
+          case avro::AVRO_BYTES: {
+            const std::vector<uint8_t>& field_value =
+                field.value<std::vector<uint8_t>>();
+            string v;
+            if (field_value.size() > 0) {
+              v.resize(field_value.size());
+              memcpy(&v[0], &field_value[0], field_value.size());
+            }
+            value[i]->flat<string>()(entry_index) = std::move(v);
+          } break;
+          case avro::AVRO_FIXED: {
+            const std::vector<uint8_t>& field_value =
+                field.value<avro::GenericFixed>().value();
+            string v;
+            if (field_value.size() > 0) {
+              v.resize(field_value.size());
+              memcpy(&v[0], &field_value[0], field_value.size());
+            }
+            value[i]->flat<string>()(entry_index) = std::move(v);
+          } break;
+          case avro::AVRO_ENUM:
+            value[i]->flat<string>()(entry_index) =
+                field.value<avro::GenericEnum>().symbol();
+            break;
+          default:
+            OP_REQUIRES(context, false,
+                        errors::InvalidArgument("unsupported data type: ",
+                                                field.type()));
+        }
+      }
+    }
+  }
+
+ private:
+  mutable mutex mu_;
+  Env* env_ GUARDED_BY(mu_);
+};
+
+class EncodeAvroOp : public OpKernel {
+ public:
+  explicit EncodeAvroOp(OpKernelConstruction* context) : OpKernel(context) {
+    OP_REQUIRES_OK(context, context->GetAttr("schema", &schema_));
+  }
+
+  void Compute(OpKernelContext* context) override {
+    // Make sure input have the same elements;
+    for (int64 i = 1; i < context->num_inputs(); i++) {
+      OP_REQUIRES(
+          context,
+          (context->input(0).NumElements() == context->input(i).NumElements()),
+          errors::InvalidArgument("number of elements different: input 0 (",
+                                  context->input(0).NumElements(),
+                                  ") vs. input ", i, " (",
+                                  context->input(i).NumElements(), ")"));
+    }
+
+    avro::ValidSchema avro_schema;
+    std::istringstream ss(schema_);
+    string error;
+    OP_REQUIRES(context, (avro::compileJsonSchema(ss, avro_schema, error)),
+                errors::Unimplemented("Avro schema error: ", error));
+
+    Tensor* output_tensor = nullptr;
+    OP_REQUIRES_OK(context, context->allocate_output(
+                                0, context->input(0).shape(), &output_tensor));
+
+    for (int64 entry_index = 0; entry_index < context->input(0).NumElements();
+         entry_index++) {
+      std::ostringstream ss;
+      std::unique_ptr<avro::OutputStream> out = avro::ostreamOutputStream(ss);
+
+      avro::GenericDatum datum(avro_schema);
+      avro::GenericRecord& record = datum.value<avro::GenericRecord>();
+      for (int i = 0; i < avro_schema.root()->names(); i++) {
+        switch (record.fieldAt(i).type()) {
+          case avro::AVRO_BOOL:
+            record.setFieldAt(i,
+                              avro::GenericDatum(
+                                  context->input(i).flat<bool>()(entry_index)));
+            break;
+          case avro::AVRO_INT:
+            record.setFieldAt(
+                i, avro::GenericDatum(static_cast<int32_t>(
+                       context->input(i).flat<int32>()(entry_index))));
+            break;
+          case avro::AVRO_LONG:
+            record.setFieldAt(
+                i, avro::GenericDatum(static_cast<int64_t>(
+                       context->input(i).flat<int64>()(entry_index))));
+            break;
+          case avro::AVRO_FLOAT:
+            record.setFieldAt(
+                i, avro::GenericDatum(
+                       context->input(i).flat<float>()(entry_index)));
+            break;
+          case avro::AVRO_DOUBLE:
+            record.setFieldAt(
+                i, avro::GenericDatum(
+                       context->input(i).flat<double>()(entry_index)));
+            break;
+          case avro::AVRO_STRING: {
+            // make a concrete explicit copy as otherwise avro may override the
+            // underlying buffer?? happens in decode (not verified in encode
+            // yet).
+            const string& v = context->input(i).flat<string>()(entry_index);
+            string field_value;
+            field_value.resize(v.size());
+            if (field_value.size() > 0) {
+              memcpy(&field_value[0], &v[0], field_value.size());
+            }
+            record.setFieldAt(i, avro::GenericDatum(field_value));
+          } break;
+          case avro::AVRO_BYTES: {
+            const string& v = context->input(i).flat<string>()(entry_index);
+            std::vector<uint8_t> field_value;
+            field_value.resize(v.size());
+            if (field_value.size() > 0) {
+              memcpy(&field_value[0], &v[0], field_value.size());
+            }
+            record.setFieldAt(i, avro::GenericDatum(field_value));
+          } break;
+          case avro::AVRO_FIXED:
+          case avro::AVRO_ENUM:
+          default:
+            OP_REQUIRES(context, false,
+                        errors::InvalidArgument("unsupported data type: ",
+                                                record.fieldAt(i).type()));
+        }
+      }
+      avro::EncoderPtr e = avro::binaryEncoder();
+      e->init(*out);
+      avro::encode(*e, datum);
+      out->flush();
+      output_tensor->flat<string>()(entry_index) = ss.str();
+    }
+  }
+
+ private:
+  string schema_;
+};
+class DecodeAvroInitOp : public ResourceOpKernel<DecodeAvroResource> {
+ public:
+  explicit DecodeAvroInitOp(OpKernelConstruction* context)
+      : ResourceOpKernel<DecodeAvroResource>(context) {
+    env_ = context->env();
+  }
+
+ private:
+  void Compute(OpKernelContext* context) override {
+    ResourceOpKernel<DecodeAvroResource>::Compute(context);
+
+    const Tensor* input_tensor;
+    OP_REQUIRES_OK(context, context->input("input", &input_tensor));
+
+    OP_REQUIRES_OK(context, resource_->Init(input_tensor->scalar<string>()()));
+  }
+  Status CreateResource(DecodeAvroResource** resource)
+      EXCLUSIVE_LOCKS_REQUIRED(mu_) override {
+    *resource = new DecodeAvroResource(env_);
+    return Status::OK();
+  }
+
+ private:
+  mutable mutex mu_;
+  Env* env_ GUARDED_BY(mu_);
+};
+
 REGISTER_KERNEL_BUILDER(Name("IO>KafkaDataset").Device(DEVICE_CPU),
                         KafkaDatasetOp);
 
@@ -751,4 +1385,17 @@ REGISTER_KERNEL_BUILDER(
 
 REGISTER_KERNEL_BUILDER(Name("IO>KafkaOutputSequenceFlush").Device(DEVICE_CPU),
                         OutputSequenceFlushOp<KafkaOutputSequence>);
+REGISTER_KERNEL_BUILDER(Name("IO>KafkaReadableInit").Device(DEVICE_CPU),
+                        IOInterfaceInitOp<KafkaReadable>);
+REGISTER_KERNEL_BUILDER(Name("IO>KafkaReadableRead").Device(DEVICE_CPU),
+                        IOReadableReadOp<KafkaReadable>);
+
+REGISTER_KERNEL_BUILDER(Name("IO>KafkaDecodeAvro").Device(DEVICE_CPU),
+                        DecodeAvroOp);
+REGISTER_KERNEL_BUILDER(Name("IO>KafkaEncodeAvro").Device(DEVICE_CPU),
+                        EncodeAvroOp);
+REGISTER_KERNEL_BUILDER(Name("IO>KafkaDecodeAvroInit").Device(DEVICE_CPU),
+                        DecodeAvroInitOp);
+
+}  // namespace data
 }  // namespace tensorflow
